requirements.txt
streamlit==1.32.2
pandas
numpy
scikit-learn
autogluon.tabular
